Reorganize channel software services
