+++ java/core/src/main/java/com/redhat/rhn/frontend/xmlrpc/channel/software/ChannelSoftwareHandler.java
@@ -88,19 +88,19 @@ import com.redhat.rhn.manager.user.UserManager;
 import com.redhat.rhn.taskomatic.TaskomaticApi;
 import com.redhat.rhn.taskomatic.task.errata.ErrataCacheWorker;
 
-import com.suse.impl.channel.software.SyncFromOriginalServiceImpl;
-import com.suse.impl.channel.software.SyncFromSourceServiceImpl;
-import com.suse.impl.channel.software.SyncFromVendorServiceImpl;
 import com.suse.manager.api.ApiIgnore;
 import com.suse.manager.api.ReadOnly;
+import com.suse.manager.channel.software.SyncFromOriginalService;
+import com.suse.manager.channel.software.SyncFromOriginalServiceImpl;
+import com.suse.manager.channel.software.SyncFromSourceService;
+import com.suse.manager.channel.software.SyncFromSourceServiceImpl;
+import com.suse.manager.channel.software.SyncFromVendorService;
+import com.suse.manager.channel.software.SyncFromVendorServiceImpl;
+import com.suse.manager.channel.software.dto.ErrataCriteria;
+import com.suse.manager.channel.software.dto.SyncOperation;
+import com.suse.manager.channel.software.dto.SyncRequest;
+import com.suse.manager.channel.software.dto.SyncResponse;
 import com.suse.manager.webui.services.pillar.MinionPillarManager;
-import com.suse.spec.channel.software.SyncFromOriginalService;
-import com.suse.spec.channel.software.SyncFromSourceService;
-import com.suse.spec.channel.software.SyncFromVendorService;
-import com.suse.spec.channel.software.dto.ErrataCriteria;
-import com.suse.spec.channel.software.dto.SyncOperation;
-import com.suse.spec.channel.software.dto.SyncRequest;
-import com.suse.spec.channel.software.dto.SyncResponse;
 
 import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
