+++ java/core/src/main/java/com/suse/manager/channel/software/SyncFromOriginalServiceImpl.java
@@ -8,7 +8,7 @@
  * along with this software; if not, see
  * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
  */
-package com.suse.impl.channel.software;
+package com.suse.manager.channel.software;
 
 import static com.redhat.rhn.common.ErrorReportingStrategies.logReportingStrategy;
 import static com.redhat.rhn.common.ErrorReportingStrategies.rpcValidationReportingStrategy;
@@ -28,10 +28,9 @@ import com.redhat.rhn.frontend.events.SyncFromOriginalErrataEvent;
 import com.redhat.rhn.manager.channel.ChannelManager;
 import com.redhat.rhn.manager.errata.ErrataManager;
 
-import com.suse.impl.channel.software.helper.ErrataResolver;
-import com.suse.spec.channel.software.SyncFromOriginalService;
-import com.suse.spec.channel.software.dto.SyncRequest;
-import com.suse.spec.channel.software.dto.SyncResponse;
+import com.suse.manager.channel.software.dto.SyncRequest;
+import com.suse.manager.channel.software.dto.SyncResponse;
+import com.suse.manager.channel.software.helper.ErrataResolver;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
