+++ java/core/src/test/java/com/suse/manager/channel/software/ChannelSoftwareTestUtils.java
@@ -8,13 +8,13 @@
  * along with this software; if not, see
  * http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
  */
-package com.suse.impl.channel.software;
+package com.suse.manager.channel.software;
 
 import static com.redhat.rhn.common.ExceptionMessage.NOT_INSTANTIABLE;
 
-import com.suse.spec.channel.software.dto.ErrataCriteria;
-import com.suse.spec.channel.software.dto.SyncOperation;
-import com.suse.spec.channel.software.dto.SyncRequest;
+import com.suse.manager.channel.software.dto.ErrataCriteria;
+import com.suse.manager.channel.software.dto.SyncOperation;
+import com.suse.manager.channel.software.dto.SyncRequest;
 
 import java.util.Arrays;
